requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
grpcio==1.69.0
grpcio-reflection==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
idna==3.10
protobuf==5.29.2
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
